Extracted common functionality from compontents
