+++ autoscheduler/frontend/src/components/SchedulingPage/CourseSelectColumn/CourseSelectCard/ExpandedCourseCard/SectionSelect/SectionInfo.tsx
@@ -7,8 +7,12 @@ import { useDispatch } from 'react-redux';
 import { toggleSelected } from '../../../../../../redux/actions/courseCards';
 import { SectionSelected } from '../../../../../../types/CourseCardOptions';
 import Meeting, { MeetingType, MeetingTypeDescription } from '../../../../../../types/Meeting';
-import { formatTime } from '../../../../../../utils/timeUtil';
+import formatMeetingDays from '../../../../../../utils/formatMeetingDays';
+import meetingBuilding from '../../../../../../utils/meetingBuilding';
+import meetingsForSection from '../../../../../../utils/meetingsForSection';
+import meetingTimeText from '../../../../../../utils/meetingTimeText';
 import GradeDist from './GradeDist/GradeDist';
+import MeetingTypeDisplay from './MeetingType/MeetingTypeDisplay';
 import * as styles from './SectionSelect.css';
 
 interface SectionInfoProps {
@@ -44,57 +48,6 @@ const SectionInfo: React.FC<SectionInfoProps> = ({
     )
     : null;
 
-  const formatMeetingDays = (meeting: Meeting): string => {
-    const DAYS_OF_WEEK = ['M', 'T', 'W', 'R', 'F', 'S', 'U'];
-    return meeting.meetingDays.reduce((acc, curr, idx) => (curr ? acc + DAYS_OF_WEEK[idx] : acc), '');
-  };
-
-  const getMeetingTimeText = (mtg: Meeting): string => {
-    if (mtg.startTimeHours === 0) {
-      // If the time is 00:00, then it's meeting time is not applicable
-      return 'N/A';
-    }
-
-    // Returns it in the format 12:00 - 1:00
-    return `${formatTime(mtg.startTimeHours, mtg.startTimeMinutes)}
-        - ${formatTime(mtg.endTimeHours, mtg.endTimeMinutes)}`;
-  };
-
-  /**
-     * Accepts an array of meetings and returns a filtered array without duplicate meetings.
-     * Meetings are considered to be duplicates if they are of the same type, meet on the same days,
-     * and start at the same time. Meetings that are the same by all of these criteria but
-     * differ only in the end times will still be considered duplicates
-     * @param arr
-     */
-  const filterDuplicateMeetings = (arr: Meeting[]): Meeting[] => {
-    // helper function to merge two meetings
-    const mergeMeetings = (mtg1: Meeting, mtg2: Meeting): Meeting => {
-      if (!mtg2) return mtg1;
-
-      // choose the later end time
-      const [laterEndHours, laterEndMinutes] = mtg2.endTimeHours > mtg1.endTimeHours
-        ? [mtg2.endTimeHours, mtg2.endTimeMinutes]
-        : [mtg1.endTimeHours, mtg1.endTimeMinutes];
-        // merge the days array by logical OR of each element
-      const days = mtg1.meetingDays.map((hasMeeting, idx) => hasMeeting || mtg2.meetingDays[idx]);
-      return {
-        ...mtg1,
-        endTimeHours: laterEndHours,
-        endTimeMinutes: laterEndMinutes,
-        meetingDays: days,
-      };
-    };
-
-    // add all meetings to a map, then get the values of the map
-    const uniqueMeetings = new Map<string, Meeting>();
-    arr.forEach((mtg) => {
-      const key = `${mtg.meetingType}${mtg.startTimeHours}${mtg.startTimeMinutes}`;
-      uniqueMeetings.set(key, mergeMeetings(mtg, uniqueMeetings.get(key)));
-    });
-    return [...uniqueMeetings.values()];
-  };
-
   // builds a div containing the section's number and available/max enrollment
   const remainingSeats = section.maxEnrollment - section.currentEnrollment;
   const remainingSeatsColor = remainingSeats > 0 ? 'black' : 'red';
@@ -113,36 +66,24 @@ const SectionInfo: React.FC<SectionInfoProps> = ({
     </Typography>
   );
 
-  // adds a tooltip for meeting types that aren't very obvious ex -> INS, PRL, etc.
-  const formatMeetingType = (mtg: Meeting): JSX.Element | string => {
-    const meetingTypeDescription = MeetingTypeDescription.get(mtg.meetingType);
-    if (meetingTypeDescription) {
-      return (
-        <Tooltip title={meetingTypeDescription} arrow placement="bottom" PopperProps={{ disablePortal: true }}>
-          <span className={styles.meetingType}>{MeetingType[mtg.meetingType]}</span>
-        </Tooltip>
-      );
-    }
-    return MeetingType[mtg.meetingType];
-  };
-
   const renderMeeting = (mtg: Meeting, showSectionNum: boolean): JSX.Element => (
     <React.Fragment key={mtg.id}>
       {showSectionNum ? sectionHeader : null}
       <Typography className={styles.denseListItem} color="textSecondary" component="tr">
-        <td>{formatMeetingType(mtg)}</td>
-        <td>{mtg.building || 'ONLINE'}</td>
+        <td>
+          <MeetingTypeDisplay meeting={mtg} />
+        </td>
+        <td>{meetingBuilding(mtg)}</td>
         <td>{formatMeetingDays(mtg)}</td>
-        <td>{getMeetingTimeText(mtg)}</td>
+        <td>{meetingTimeText(mtg)}</td>
       </Typography>
     </React.Fragment>
   );
 
   // filters and then builds UI elements for the meetings that match this section
-  const meetingRows = filterDuplicateMeetings(
-    meetings.filter((mtg) => mtg.section.id === section.id),
-  ).map((mtg, mtgIdx) => renderMeeting(mtg, mtgIdx === 0));
-
+  const meetingRows = meetingsForSection(section, meetings).map((mtg, mtgIdx) => (
+    renderMeeting(mtg, mtgIdx === 0)
+  ));
 
   // makes a list of the meetings in this section, along with one checkbox for all of them
   const sectionDetails = (
