+++ autoscheduler/frontend/src/utils/meetingBuilding.ts
@@ -1,0 +1,5 @@
+import Meeting from "../types/Meeting";
+
+export default function meetingBuilding(meeting: Meeting): string {
+  return meeting.building || 'ONLINE';
+}
